requirements.txt
pylint
pytest
flask
snowflake-connector-python== 4.1.0

